# [SMALLFIX] Improvement to openAtPosition when delegation is turned on, for S3/GCS.