--- a/core/server/src/main/java/alluxio/worker/file/UnderFileSystemManager.java
+++ b/core/server/src/main/java/alluxio/worker/file/UnderFileSystemManager.java
@@ -20,6 +20,7 @@ import alluxio.exception.FileAlreadyExistsException;
 import alluxio.exception.FileDoesNotExistException;
 import alluxio.exception.PreconditionMessage;
 import alluxio.underfs.UnderFileSystem;
+import alluxio.underfs.s3.S3UnderFileSystem;
 import alluxio.util.IdUtils;
 import alluxio.util.io.PathUtils;
 import alluxio.util.network.NetworkAddressUtils;
@@ -165,6 +166,10 @@ public final class UnderFileSystemManager {
         return null;
       }
       UnderFileSystem ufs = UnderFileSystem.get(mUri, mConfiguration);
+      // Special handing for S3 and GCS to save redundant ufs open and close in S3/GCS skip.
+      if (ufs instanceof S3UnderFileSystem) {
+        return ((S3UnderFileSystem) ufs).openAtPosition(mUri, position);
+      }
       InputStream stream = ufs.open(mUri);
       if (position != stream.skip(position)) {
         throw new IOException(ExceptionMessage.FAILED_SKIP.getMessage(position));