--- a/underfs/gcs/src/main/java/alluxio/underfs/gcs/GCSInputStream.java
+++ b/underfs/gcs/src/main/java/alluxio/underfs/gcs/GCSInputStream.java
@@ -63,6 +63,25 @@ public final class GCSInputStream extends InputStream {
     mInputStream = new BufferedInputStream(mObject.getDataInputStream());
   }
 
+  /**
+   * Creates a new instance of {@link GCSInputStream}, at a specific position.
+   *
+   * @param bucketName the name of the bucket
+   * @param key the key of the file
+   * @param client the client for GCS
+   * @param pos the position to start
+   * @throws ServiceException if a service exception occurs
+   */
+  GCSInputStream(String bucketName, String key, GoogleStorageService client, long pos)
+      throws ServiceException {
+    mBucketName = bucketName;
+    mKey = key;
+    mClient = client;
+    mPos = pos;
+    mObject = mClient.getObject(mBucketName, mKey, null, null, null, null, mPos, null);
+    mInputStream = new BufferedInputStream(mObject.getDataInputStream());
+  }
+
   @Override
   public void close() throws IOException {
     mInputStream.close();