--- a/underfs/gcs/src/main/java/alluxio/underfs/gcs/GCSUnderFileSystem.java
+++ b/underfs/gcs/src/main/java/alluxio/underfs/gcs/GCSUnderFileSystem.java
@@ -285,6 +285,24 @@ public class GCSUnderFileSystem extends UnderFileSystem {
     }
   }
 
+  /**
+   * Opens a GCS object at position and returns the input stream.
+   *
+   * @param path the GCS object path
+   * @param pos the position to open at
+   * @return the opened input stream
+   * @throws IOException if failed to open file at position
+   */
+  public InputStream openAtPosition(String path, long pos) throws IOException {
+    try {
+      GSObject obj = mClient.getObject(mBucketName, path, null, null, null, null, pos, null);
+      return obj.getDataInputStream();
+    } catch (ServiceException e) {
+      LOG.error("Failed to open file {} at position {}:", path, pos, e);
+      return null;
+    }
+  }
+
   @Override
   public boolean rename(String src, String dst) throws IOException {
     if (!exists(src)) {