--- a/underfs/gcs/src/main/java/alluxio/underfs/gcs/GCSUnderFileSystem.java
+++ b/underfs/gcs/src/main/java/alluxio/underfs/gcs/GCSUnderFileSystem.java
@@ -296,8 +296,7 @@ public class GCSUnderFileSystem extends UnderFileSystem {
   public InputStream openAtPosition(String path, long pos) throws IOException {
     try {
       path = stripPrefixIfPresent(path);
-      GSObject obj = mClient.getObject(mBucketName, path, null, null, null, null, pos, null);
-      return obj.getDataInputStream();
+      return new GCSInputStream(mBucketName, path, mClient, pos);
     } catch (ServiceException e) {
       LOG.error("Failed to open file {} at position {}:", path, pos, e);
       return null;