This PR includes an improvement for `openAtPosition` when delegation is turned on. Implementation of `openAtPosition` is added for S3UnderFileSystem, to save redundant close and reopen S3InputStream.

A simple benchmark was run locally, doing a Read from a 10MB S3 THROUGH file, with read buffer size = 2M. That being said, with delegation on, `openAtPosition` will be called for 5 times. The logs measure the time elapsed for `OpenAtPosition`. The data shows that this PR would save > 40% time of `OpenAtPosition` call, for each 2M chunk.

Before this PR.
```
➜  tachyon git:(branch-1.1-s3) ✗ time ./bin/alluxio fs copyToLocal /file10M.txt /tmp/file10M.txt
Copied /file10M.txt to /tmp/file10M.txt
./bin/alluxio fs copyToLocal /file10M.txt /tmp/file10M.txt
1.18s user 0.15s system 6% cpu 20.415 total

2016-06-02 13:38:01,067 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 147 (ms)
2016-06-02 13:38:01,067 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 148 (ms)
2016-06-02 13:38:04,242 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 1021 (ms)
2016-06-02 13:38:04,614 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 1393 (ms)
2016-06-02 13:38:07,532 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 373 (ms)
2016-06-02 13:38:07,923 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 764 (ms)
2016-06-02 13:38:10,655 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 433 (ms)
2016-06-02 13:38:11,042 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 820 (ms)
2016-06-02 13:38:13,636 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 392 (ms)
2016-06-02 13:38:14,033 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 789 (ms)

➜  tachyon git:(branch-1.1-s3) ✗ time ./bin/alluxio fs copyToLocal /file10M.txt /tmp/file10M.txt
Copied /file10M.txt to /tmp/file10M.txt
./bin/alluxio fs copyToLocal /file10M.txt /tmp/file10M.txt
1.16s user 0.14s system 7% cpu 17.565 total

2016-06-02 13:38:24,595 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 109 (ms)
2016-06-02 13:38:24,596 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 110 (ms)
2016-06-02 13:38:27,223 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 382 (ms)
2016-06-02 13:38:27,595 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 754 (ms)
2016-06-02 13:38:30,007 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 482 (ms)
2016-06-02 13:38:30,474 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 949 (ms)
2016-06-02 13:38:33,067 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 377 (ms)
2016-06-02 13:38:33,457 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 766 (ms)
2016-06-02 13:38:36,996 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition ufs.open took 368 (ms)
2016-06-02 13:38:37,404 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 775 (ms)
```

After this PR:
```
git:(branch-1.1-s3) ✗ time ./bin/alluxio fs copyToLocal /file10M.txt /tmp/file10M.txt
Copied /file10M.txt to /tmp/file10M.txt
./bin/alluxio fs copyToLocal /file10M.txt /tmp/file10M.txt
1.17s user 0.14s system 7% cpu 16.462 total

2016-06-02 13:35:22,860 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 222 (ms)
2016-06-02 13:35:22,860 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 222 (ms)
2016-06-02 13:35:25,622 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 487 (ms)
2016-06-02 13:35:25,623 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 487 (ms)
2016-06-02 13:35:28,143 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 376 (ms)
2016-06-02 13:35:28,143 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 376 (ms)
2016-06-02 13:35:30,708 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 384 (ms)
2016-06-02 13:35:30,709 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 384 (ms)
2016-06-02 13:35:33,786 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 397 (ms)
2016-06-02 13:35:33,786 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 397 (ms)

➜  tachyon git:(branch-1.1-s3) ✗ time ./bin/alluxio fs copyToLocal /file10M.txt /tmp/file10M.txt
Copied /file10M.txt to /tmp/file10M.txt
./bin/alluxio fs copyToLocal /file10M.txt /tmp/file10M.txt
1.15s user 0.15s system 8% cpu 15.916 total

2016-06-02 13:40:15,297 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 162 (ms)
2016-06-02 13:40:15,298 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 162 (ms)
2016-06-02 13:40:17,972 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 409 (ms)
2016-06-02 13:40:17,972 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 409 (ms)
2016-06-02 13:40:20,694 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 295 (ms)
2016-06-02 13:40:20,695 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 295 (ms)
2016-06-02 13:40:23,851 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 398 (ms)
2016-06-02 13:40:23,852 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 398 (ms)
2016-06-02 13:40:26,481 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - S3OpenAtPosition took 390 (ms)
2016-06-02 13:40:26,482 INFO  logger.type (UnderFileSystemManager.java:openAtPosition) - OpenAtPosition total took 390 (ms)
```