@ifcharming 2016-06-02
Thanks @calvinjia for the review!
I updated the stream logic to be within `S3InputStream`. PTAL